fix(candle-gen-krea): type SC#2 cos-gap band f64; add PR CUDA compile-check (sc-12110)
`nvfp4_krea_dit_gpu.rs` (introduced by #28) does not compile under `--features
cuda`: in `nvfp4_krea_dit_sc2_parity_vs_q4_tier`, `MEASURED_COS_GAP` and
`COS_GAP_TOL` were `f32` but are compared against `cos_gap`, an `f64` derived
from `cosine()` (which returns `f64`) — three E0308/E0277 errors. Declare both
constants `f64` (cosine gaps live in the f64 domain), which is the only type
mismatch in the file.

This reached main because nothing on the PR path compiled the cuda-gated
targets. The whole file is `#![cfg(feature = "cuda")]`, which the candle-cpu
lane's `--all-targets` never sees (feature off), and the `windows-cuda` lane —
which *does* build `--lib --tests --features cuda` — was `workflow_dispatch`-only
and never ran for the PR.

Add `windows-cuda-build`: a compile-only lane (`cargo test --no-run --lib
--tests --features cuda`) that runs on every PR/push selecting the CUDA lane, so
a non-compiling cuda-gated `#[ignore]`d GPU test can't merge again. `--no-run`
compiles every lib + test binary without touching the GPU, so it needs the CUDA
toolchain but not an exclusive throughput run. The full build-and-run
`windows-cuda` lane stays manual (needs an exclusive GPU). README updated to
describe the split.

+++ crates/media/candle-gen/candle-gen-krea/tests/nvfp4_krea_dit_gpu.rs
@@ -700,10 +700,12 @@ fn nvfp4_krea_dit_sc2_parity_vs_q4_tier() {
     // ==========================================================================================
     // Q4 − NVFP4-W4A16 cosine gap measured 2026-07-15 on real weights (this leg quantizes all 260
     // projections; the q4 tier leaves 4 dense — see the surface-mismatch confound in the docs).
-    const MEASURED_COS_GAP: f32 = 0.00601;
+    // f64 to match `cosine()`'s f64 domain: `cos_gap` below is an f64 difference of two f64 cosines,
+    // and this band is compared against it directly.
+    const MEASURED_COS_GAP: f64 = 0.00601;
     // Absorbs GEMM/library nondeterminism only. The seed, prompt and schedule are fixed, so this is
     // NOT a chaos budget — chaos is *between* the tiers, and it is exactly what the band pins.
-    const COS_GAP_TOL: f32 = 0.004;
+    const COS_GAP_TOL: f64 = 0.004;
 
     let cos_gap = q4_cos - w16_cos;
     assert!(
